+++ ql/Pricers/himalaya.hpp
@@ -55,7 +55,7 @@ namespace QuantLib {
                      Rate riskFreeRate,
                      double strike,
                      const std::vector<Time>& times,
-                     long samples,
+                     unsigned int samples,
                      bool antitheticVariance,
                      long seed = 0);
         };
